requirements.txt
Sphinx~=2.2.1
pysphinxdoc~=1.1.3
pandas~=1.1.4
numpy~=1.19.4
scipy~=1.5.4
scikit-learn~=0.23.2
torch~=1.7.0
PyYAML~=5.3.1
nitk~=0.1
hdbscan~=0.8.26
cvxpy~=1.1.7
matplotlib~=3.3.3
nibabel~=3.2.0
nilearn~=0.7.0
joblib~=0.17.0
statsmodels~=0.12.1
scikit-image~=0.15.0
torchvision~=0.8.1
boto3~=1.16.19
botocore~=1.19.19
h5py~=3.1.0
requests~=2.22.0
Pillow~=7.0.0
visdom~=0.1.8.9
PySide2~=5.15.1
hiddenlayer~=0.3
torchviz~=0.0.1
tabulate~=0.8.7
setuptools~=45.2.0